Move blog article row to blog sub-directory and revamp with new classes.
