+++ src/components/blog/article/row.tsx
@@ -4,8 +4,9 @@ import { type Article } from "@prisma/client";
 
 import { api } from "@utils/api";
 import SuccessBox from "@utils/success";
-import Comment from "@utils/icons/comment";
-import View from "@utils/icons/view";
+import CommentIcon from "@utils/icons/comment";
+import ViewIcon from "@utils/icons/view";
+import IconAndText from "@components/containers/icon_and_text";
 
 const Row: React.FC<{
     article: Article,
@@ -33,39 +34,41 @@ const Row: React.FC<{
                 />
             ) : (
                 <div className={"article-row " + ((small) ? "article-row-sm" : "article-row-lg")}>
-                    <div className={"w-full " + ((small) ? "h-48" : "h-72")}>
-                        <img src={banner} className="w-full h-full max-h-full" />
+                    <div className="article-row-image">
+                        <img src={banner} />
                     </div>
-                    <div className="">
-                        <h3 className="text-white text-2xl font-bold text-center">{article.title}</h3>
+                    <div className="article-row-title">
+                        <h3>{article.title}</h3>
                     </div>
-                    <div className="pb-6 grow">
-                        <p className="text-gray-100 text-sm">{article.desc}</p>
+                    <div className="article-row-description">
+                        <p>{article.desc}</p>
                     </div>
-                    <div className="pb-6 flex justify-between text-white text-sm">
-                        <div className="flex flex-wrap items-center">
-                            <span>
-                                <View
+                    <div className="article-row-stats">
+                        <IconAndText
+                            icon={
+                                <ViewIcon
                                     classes={["w-6", "h-6", "fill-white"]}
                                 />
-                            </span>
-                            <span className="ml-1">{article.views}</span>
-                        </div>
-                        <div className="flex flex-wrap items-center">
-                            <span>
-                                <Comment 
+                            }
+                            text={<>{article.views}</>}
+                            inline={true}
+                        />
+                        <IconAndText
+                            icon={
+                                <CommentIcon 
                                     classes={["w-6", "h-6", "fill-white", "stroke-white"]}
                                 />
-                            </span>
-                            <span className="ml-1">{article.comments}</span>
-                        </div>
+                            }
+                            text={<>{article.comments}</>}
+                            inline={true}
+                        />
                     </div>
-                    <div className="p-6 flex flex-wrap gap-2 justify-center">
-                        <Link className="w-full button" href={viewUrl}>Read More</Link>
+                    <div className="article-row-read-more">
+                        <Link className="button" href={viewUrl}>Read More</Link>
                     </div>
-                    <div className="p-6 flex flex-wrap gap-2 justify-center">
-                        <Link className="w-full button button-secondary" href={editUrl}>Edit</Link>
-                        <Link className="w-full button button-delete" href="#" onClick={(e) => {
+                    <div className="article-row-actions">
+                        <Link className="button button-secondary" href={editUrl}>Edit</Link>
+                        <Link className="button button-delete" href="#" onClick={(e) => {
                             e.preventDefault();
 
                             const yes = confirm("Are you sure you want to delete this article?");
