+++ src/pages/index.tsx
@@ -6,7 +6,7 @@ import { prisma } from "@server/db";
 import { type Article, type Partner, type Service, type User } from "@prisma/client";
 
 import Wrapper from "@components/wrapper";
-import ArticleRow from "@components/article/row";
+import ArticleRow from "@components/blog/article/row";
 import UserRow from "@components/user/row";
 import ServiceRow from "@components/service/row";
 import PartnerRow from "@components/partner/row";
